requirements.txt
black==23.3.0
factory-boy==3.2.1
faker==18.3.1
flake8==6.0.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
isort==5.12.0
mypy==1.3.0
pytest==7.3.1
pytest-cov==4.0.0
SQLAlchemy==2.0.15